+++ src/emdbg/debug/px4/buffer.py
@@ -3,6 +3,7 @@
 
 from __future__ import annotations
 from .base import Base
+from .utils import gdb_len
 
 
 class UartBuffer(Base):
@@ -34,3 +35,31 @@ class UartBuffer(Base):
             content = "".join(chr(v) if chr(v).isprintable() else f"\\{v:02x}"
                               for v in content.tobytes())
             yield ("content", content)
+
+
+class ConsoleBuffer(Base):
+    """
+    Pretty Printing Console buffers
+    """
+    def __init__(self, gdb, buf_ptr: "gdb.Value"):
+        super().__init__(gdb)
+        self._buf = buf_ptr
+
+    def to_string(self) -> str:
+        ptr = int(self._buf['_buffer'].address)
+        size = gdb_len(self._buf['_buffer'])
+        head, tail = self._buf['_tail'], self._buf['_head']
+        used = head - tail if (tail <= head) else size - tail + head
+        header = f"ConsoleBuffer({used}B/{size}B: "
+        if tail <= head: header += f"[{tail} -> {head}]) =\n"
+        else: header += f"{head}] <- [{tail}) =\n"
+        if used:
+            if (tail <= head):
+                # [tail, head]
+                content = self.read_memory(ptr + tail, used).tobytes()
+            else:
+                # head], [tail
+                content = self.read_memory(ptr + tail, size - tail).tobytes()
+                content += self.read_memory(ptr, head).tobytes()
+            header += "".join(chr(v) for v in content)
+        return header